Refactored into modules and moved display actions to functions
