+++ src/board.js
@@ -1,0 +1,16 @@
+const gameBoard = (() => {
+  const WINS = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [6, 4, 2],
+  ];
+
+  return { WINS };
+})();
+
+export default gameBoard;
